+++ NEWS.md
@@ -1,5 +1,11 @@
 # Changelog BBWPC
 
+## 0.4.0
+### Added
+- add wrapper function `bbwp` to run BBWP for a series of fields
+- add wrapper function `ecoregeling` to run Ecoregeling for a series of fields
+- add associated test functions for `bbwp` and `ecoregeling`
+
 ## 0.3.2
 ### Added
 - add argument `B_SLOPE` to `bbwp_field_properties`,`bbwp_field_indicators`, issue #4